+++ src/components/Section/index.ts
@@ -1,3 +1,3 @@
-export { createSection } from "./Section";
+export { Section, createSection } from "./Section";
 export type { SectionProps, SectionOverrides, SectionDataProps } from "./Section";
 export * from "./variants";